feat(bff): handle crayon code block rewrites
